Fix #58: roll back the transaction on error
The REST transactional endpoint will always return a transaction, but that
transaction will actually have been rolled back server-side if the response
contains errors. The transaction is now also considered rolled back
client-side in that case.

It allows the Connection to be closed properly, because it won't try to
roll back the transaction explicitly, which fails and interrupts the closing
when the transaction does not exist anymore.