Implement more methods in ResultSetImpl (#60)
In particular, this update adds implementation to the remaining column getter methods and also methods like `wasNull`, `findColumn`, `unwrap` and `isWrapperFor`.

In addition, this fixes a bug where getters by column label would throw an exception stating that an existing label is absent.

The `getWarnings` method is not currently supported. The `getMetaData` should be implemented separately.