refactor: Make SQL parser not fail hard on unknown functions. (#80)
While we do cover a lot of functions explicitly, we haven't covered all and some Cypher functions might just not exist in SQL.
This PR tells jOOQ to ignore unknown functions.
They will endup as custom functions in the QOM which we then can translate into our custom functions.

Additionally the PR changed the behaviour of name parsing to `AS_IS`. Normally I'd consider this breaking, but there is no release yet.
`AS_IS` applies to all names (tables, views and especially functions).
Thus, making sure we would call the function the user wishes for or put in another way: We delegate to the user dealing with the fact that in Cypher some functions are case sensitive (all user defined ones) and most built-in are not.