feat: Implement transaction management on connection level.
As part of this work, transaction management is implemented. Generally speaking, the connection is responsible for transaction management, but individual transaction logic is delegated to the internal `Transaction` implementation that is also responsible for Bolt connection interaction in regards to the given transaction.

Database metadata will be fetched within the existing transaction if one is available or will initiate a new transaction that gets rolled back when metadata is fetched.

The `Bolt` `RESET` message is pipelined with `BEGIN` if there was an applicable error that might be reset upon next exchange.