refactor: Extract all metadata queries used into a separate resource.
Thus we avoid the (nice) JDK 17 string literals and (ugly) String concatenations.
In addition it could open up the possibility to have several variants of those, for past and future Neo4j versions.

Properties format has been chosen because it's one of the easiest to read without dependencies.
XML would be another option, but that would actually but more weight in on GraalVM native image or the module path.

Adding libraries for JSON or YAML is not worth the additional weight for this.