feat: Add support for chaining SQL translators.
SQL translators now can be ordered and you can have as many SQL translator factories as you want on the class- or module-path. The will be called in order of highest precedence, and individual translators but the last are free to fail or to modify the intermediate statement.

No additional translators are provided in this commit.

Co-authored-by: Gerrit Meier <meistermeier@gmail.com>