feat: Implement Connection Network Timeout. (#564)
The timeout value applies to individual server responses, not the JDBC API calls. Any response from the server, including the `NOOP` chunk, interrupts the idle period.

When initializing a new Bolt connection, the Neo4j server may supply a [connection.recv_timeout_seconds](https://neo4j.com/docs/bolt/current/appendix/connection-hints/#hint-recv-timeout-seconds) connection hint that defines the amount of time for which the connection may remain idle. When the network timeout value is set to `0` and the hint value is available, the latter is used as a default.

This update also fixes an issue when database name would not be supplied in the `BEGIN` message for explicit transaction. Also, the commit handler has been updated to handle a missing bookmark.

Closes #233.