refactor: Rework catalog support.
This commit changes the support for catalogs more towards the JDBC spec.

We will align with the concept of

```
DBMS -< 1..* Catalogs (Databases) -< 1 Schema (Always public)
```

The catalog of a connection is equal to the database being connected to and cannot be changed for the time being.

All metadata queries will work with a `null`, blank or catalog parameter equals to the current database now.
All metadata items will report their catalog now being equal to the database name.

Signed-off-by: Michael Simons <michael.simons@neo4j.com>