fix: Allow matching on multiple patterns, check target nodes in relationships. [translator]
This change allows first and foremost that the translator can now not only match on chains of relationships, but also on multiple, indirectly connected patterns.

This sets the groundwork to be able to create an equivalent match for a chain of join statements that are not really a chain, but starts with a driving table that joins to several other tables.

Fixes #814.

Signed-off-by: Michael Simons <michael.simons@neo4j.com>