feat: Clean up mixed SQL/Cypher queries generated by Spark. [translator]
This adds a new translator implementation, that by default is always running before the default implementation. It will check if a query looks like something that Spark did.

If this is the case, the idea is to check whether the inner query is a already a cypher query. If that’s the case, no further translation will be attempted.