feat(metadata): Use `apoc.meta.schema` if available for computing tables and relationships.
Falls back to the original query if apoc is not installed.
Introduced an internal helper on the metadata to check for existince of apoc in case we will have more of this.
Just calling `apoc.version` for a check does not cut it, as it will invalidate any ongoing transaction and works only on a fresh meta data instance.