refactor: Replace several more unnessary usages of `SQLFeatureNotSupportedException`.
This changes especially changes the way the local statement is connected to the result set produced from meta data: Both are now connected, so that the result set is more standard conform.

Other low hanging fruits incldue the type map and the prepared statement without generated keys in the connection implementation. The latter is always supported anyhow, so we might as well just check if the type requested fits and can return it.

Signed-off-by: Michael Simons <michael.simons@neo4j.com>