refactor(metadata): Improve transaction handling in metadata.
Decoupling the connection and the transaction in the metadata just to circumvent the auto commit check is brittle and errorprone, as tests with Flyway have shown.

We must however still be able to rollback by default, as commands like `SHOW USER` cannot run in the same transactions like normal queries due to restrictions in Neo4j.

So it is best to use the actual existing JDBC auto commit machinery for the purpose.