refactor(metadata): Improve transaction handling in metadata. (#877)
Decoupling the connection and the transaction in the metadata just to circumvent the auto commit check is brittle and errorprone, as tests with Flyway have shown.

We must however still be able to rollback by default, as commands like `SHOW USER` cannot run in the same transactions like normal queries due to restrictions in Neo4j.

So it is best to use the actual existing JDBC auto commit machinery for the purpose. For that however we must be more lenient with rollbacks and commits: There is no need to fail a rollback or commit when there is no actual ongoing transaction, nothing bad will happen and we make commit and rollback idempotent when any existing transaction is already committed or rolled back.