feat: Add events and metrics to the driver. (#895)
This changes adds both events and metrics to the driver.

The event system provides access to

- connections being opend and closed
- statements being created and closed
- execution results
- cache events

Listeners can be added by using `unwrap` on the driver, connection or statement classes and adding the appropriate listeners.

Build on the event system are metrics.

Metrics require Micrometer on the classpath and will configure itself to use the global registry if available. No other configuration is necessary.
Events are translated into corresponding metrics.

---------

Co-authored-by: Dmitriy Tverdiakov <11927660+injectives@users.noreply.github.com>