feat: Add support for tracing. (#904)
The core topic of this commit is to support tracing of query execution and iteration of result sets in such a way, that possible parent spans are detected and the JDBC driver will join them.

First, this change moves a bit around in the event module:

* All event methods now start with `on`
* All events are now inner records to keep the package surface small
* Some events got new attributes

This hadn’t been released prior, so no breaking changes there.

Then, a new exported package is introduced in the core module, named `tracing`. It defines a façade over any tracing api.
The façade is implemented once, based on Micrometer tracing. If something is missing, we need to change or similar, this can be added here.
In contrast to the metrics feature, we cannot get away with zero configuration. As we want no foreign API types on our core api, we delegated this to a separate module, asking the user to add this module, when they want tracing.

Tracing itself is fully build on the event system, with additional new events on the result set.

---------

Co-authored-by: Stefano Ottolenghi <stefano.ottolenghi@neo4j.com>