feat: Support Cypher-backed views. (#946)
This change introduces Cypher-backed views, a way to statically define Cypher queries that are recognized as read-only views on the client side of an application.

The feature is tied closely to the standard implementation of the SQL to Cypher translator, as this is the only place that gives us proper information if a user is actually querying a view.
Hence, they only work when SQL To Cypher translation is enabled and are not available in a Cypher query.

The change adds to the translator SPI in that regard that all implementors must report the views to the driver, so that the driver itself then can use those information in the database meta data.

---------

Signed-off-by: Michael Simons <michael.simons@neo4j.com>
Co-authored-by: Gerrit Meier <meistermeier@gmail.com>