feat: Add support for refreshing expiring token based authentication.
This adds a machinery to use `Supplier<Authentication>` with the Neo4j JDBC driver. Those suppliers can either provide non-expiring, static authentications or token based expiring authentications.

The driver will automatically manage the state of the token and ask for new tokens once the expire. Token state will be checked on the beginning of each transaction and if a token is expired, the driver will add additional bolt messages to the stream of messages reauthenticating any given connection.

In the progress of this feature

* the driver builder have been enhanced
* permutation tests for the builder have been added
* some methods have been deprecated to signal a change in the return type in the next major
* some methods have been deprecated for removal with proper replacement to align both tracing and this new feature

For tracing and the authentication supplier there are now static `registerXXX` methods on the driver class modifying the global registered instance so the is no longer a need to do this via a JDBC datasource.

The tests for this feature uses Keycloak, brought via https://github.com/dasniko/testcontainers-keycloak. The latter is a great test container module, but brings shrinkwrap and some Quarkus classes, that will mess up native image tests, hence the integration tests life in a separate Maven module.

Signed-off-by: Michael Simons <michael.simons@neo4j.com>