feat: Upgrade bolt-connection to 6.0.2 and add support for JDBC over HTTP. (#979)
This change upgrades the bolt-connection API to 6.0.2. Since 4.x providers of bolt messages are discoverable via ServiceLoader and available transports are binary bolt and new, http against Neo4j 2025.06 and higher. The way the Neo4j JDBC driver discovers bolt connection providers and initialises them has been adapted, URL parsing has been enhanced so that in addition to `jdbc:neo4j://` a subprotocol of either `http` or `https` is supported: `jdbc:neo4j:https://`. There are no further user facing changes in this feature.

Bigger changes had to be made how the bundles are created. We opted to always include the http transport so that people who want to try this out don't have to add additional dependencies. However, that requires for the full bundle a more elaborate shading.
While working on that we realised that the full bundles did not work on GraalVM native image. It would have been easy to say "this is not their intended use case", but that would create another of those scenarios in which we need to tell people  "you can do x only with setup 1, but not with setup 2 and vice versa", so we fixed that too.

Last but not least, in this change we wanted to benefit from JUnit 5.13 parameterised classes, but we learned that JUnit 5.13 does not play nicely with native-maven-build tools for GraalVM, hence, we needed to apply a workaround until there's a fix available for that.

With the workaround in place we have parameterised 

* `ConnectionIT`
* `AbstractDatabaseMetadata`

integration tests and added a pure HTTP test in the full bundle.

Additionally, we changed the build matrix to run integration tests now against last 5.26 gen and latest 2025.x version of Neo4j.

---------

Signed-off-by: Michael Simons <michael.simons@neo4j.com>
Co-authored-by: Dmitriy Tverdiakov <11927660+injectives@users.noreply.github.com>