feat: Allow manipulation of relationships via a single target table. (#1143)
This change enables the use of

- `INSERT INTO`
- `DELETE FROM`
- `TRUNCATE`
- `UPDATE`

clauses with virtual tables of type `RELATIONSHIP` as determined from the database metadata when available and in case such relationships exists as "templates", or inferred from the target table name. The latter comes with restrictions, as the translator has limited clues to determine which property is to be bound to which graph element. 

In general, the recommend way to work with this change is to either define one relationship as template like this:

```cypher
CREATE (:Start {p1: 'v1'}) -[:REL {p2: 'v2'}]->(:End {p3: 'v3'})
```

which will be represented as virtual table that can be manipulated through the table name `Start_REL_End` in such a way that for example


```sql
INSERT INTO Start_REL_END(p1, p2, p3) VALUES('v12', 'v22', 'v23')
``` 

will work as expected. If no such template is present or possible. columns can be qualified like this:

```sql
INSERT INTO Start_REL_END(Start.p1, REL.p2, END.p3) VALUES('v12', 'v22', 'v23')
``` 

---------
Signed-off-by: Michael Simons <michael.simons@neo4j.com>