fix: Improve relationship insertation.
This fix tackles various issues with the first iteration of relationship insertion:

Columns must only be ever used _once_ for insertation:
* If a column is qualified and the qualifier matches the target label, it is used for that node or relationship
* If it is unqualified it may be only used if it is in the existing properties for the entities or it has not been used so far

That means we must keep track about the used columns, and those cannot be tracked by their name alone, they must be tracked by their qualified name

To make this work with template relationship, virtual tables must include the properties of start and end node. They are now included with a SCOPED_TABLE that is not null. This also needed alignment with SELECT *.

Also, some tools decided to repeat the virtual id columns during inseration. They must be tackled individually and we now use them as a where node predicate for matching start and end.

This change closes #1155.

Signed-off-by: Michael Simons <michael.simons@neo4j.com>