fix: Parameter name generation and handling with index like parameters.
The translator treats $n as indexes when n is an integer value and tries to incorporate that with plain indexed parameters. jOOQ https://github.com/jOOQ/jOOQ/issues/16332 has been fixed a while ago, so our algorithm was dated. Moreover, it didn’t produce desirable results when mixing non-numeric names with numeric names, out of order numeric names. The test we had looked for a sequence, but that sequence is not meaning full in terms of parameters specificed. The test has beed adapted, and the rest has been fixed and greatly simplified.

Whether it is sensible to allow the usage of such wild mixtures is a point of a discussion, however, it is hard to prevent users, hence this attempt to make them all work.

Signed-off-by: Michael Simons <michael.simons@neo4j.com>