feat(translator): Add support for `LEFT JOIN` and improve flattening of joins. (#1170)
This add support for explicit left joins and also improves the flattening of joins, that is: If there’s enough probability that the start or end of relationship table is matched by a node table given the existing schema columns, the join is flattened.

In terms of columns returned: Columns projected from a relationship table will now always be alias, so that they don’t conflict with columns projected from the node tables.

Signed-off-by: Michael Simons <michael.simons@neo4j.com>