feat: Add full support for `PreparedStatement#setObject(int, java.lang.Object, int)` and corresponding `getObject` in result set.
Until this change the former throw `SQLFeatureNotSupportedException` and the latter worked more on a best effort basis. 

Taking the SQL Type parameter into account we can be quite precise on mapping

* Number types
* Dates

and at what Neo4j Cypher-Type we are aiming. The implementation itself might be a bit under engineered, but for now the huge if/else chain works well enough for the type that Neo4j can actually store.

Signed-off-by: Michael Simons <michael.simons@neo4j.com>