build(deps): Bump org.mockito:mockito-bom from 5.20.0 to 5.21.0 (#1217)
This update of Mockito requires a change in testing due to a change of behaviour in Mockito 5.21, which now returns non-null, completed futures for mocks, see https://github.com/mockito/mockito/pull/3727/files

We relied on those implicicity being null, so that a failure happens. When this is no longer the case, the test just hangs.

Now we are explicit on the method in question by actually returning a failed future.

Signed-off-by: Michael Simons <michael.simons@neo4j.com>